requirements.txt
pynput
pywin32
psutil
pystray
Pillow
schedule
win10toast
